[REFACTOR] Generalize client directory and shard connectors
