+++ pandora-client-web/test/components/gameContext/directoryConnectorContextProvider.test.ts
@@ -1,4 +1,3 @@
-import { act } from 'react';
 import { RenderHookResult } from '@testing-library/react';
 import { IDirectoryClientChangeEvents } from 'pandora-common';
 import {
@@ -7,7 +6,7 @@ import {
 	useDirectoryChangeListener,
 	useDirectoryConnector,
 } from '../../../src/components/gameContext/directoryConnectorContextProvider';
-import { MockAccountInfo, MockAuthToken, MockDirectoryConnector } from '../../mocks/networking/mockDirectoryConnector';
+import { DirectoryConnector } from '../../../src/networking/directoryConnector';
 import { ProvidersProps, RenderHookWithProviders } from '../../testUtils';
 
 const directoryChangeEvents: IDirectoryClientChangeEvents[] = [
@@ -17,10 +16,10 @@ const directoryChangeEvents: IDirectoryClientChangeEvents[] = [
 ];
 
 describe('DirectoryConnectorContextProvider', () => {
-	let directoryConnector: MockDirectoryConnector;
+	let directoryConnector: DirectoryConnector;
 
 	beforeEach(() => {
-		directoryConnector = new MockDirectoryConnector();
+		directoryConnector = new DirectoryConnector();
 	});
 
 	describe('useDirectoryConnector', () => {
@@ -49,67 +48,19 @@ describe('DirectoryConnectorContextProvider', () => {
 				renderHookWithTestProviders(() => useDirectoryChangeListener(event, callback, false));
 				expect(callback).not.toHaveBeenCalled();
 			});
-
-		it.each(directoryChangeEvents)('should fire the callback only when a %p event is emitted', (listenerEvent) => {
-			renderHookWithTestProviders(() => useDirectoryChangeListener(listenerEvent, callback, false));
-			expect(callback).not.toHaveBeenCalled();
-
-			let expectedCallCount = 0;
-			for (const event of directoryChangeEvents) {
-				act(() => directoryConnector.changeEventEmitter.fireEvent(event, true));
-				if (event === listenerEvent) {
-					expectedCallCount++;
-				}
-				expect(callback).toHaveBeenCalledTimes(expectedCallCount);
-			}
-		});
 	});
 
 	describe('useCurrentAccount', () => {
 		it('should return the directory connector\'s current account information', () => {
-			const { result, rerender } = renderHookWithTestProviders(useCurrentAccount);
+			const { result } = renderHookWithTestProviders(useCurrentAccount);
 			expect(result.current).toBe(directoryConnector.currentAccount.value);
-
-			const newAccountInfo = MockAccountInfo({
-				id: 999,
-				username: 'useCurrentAccountTest',
-				created: 1654862108851,
-			});
-			act(() => {
-				directoryConnector.currentAccount.value = newAccountInfo;
-			});
-			rerender();
-			expect(result.current).toBe(newAccountInfo);
-
-			act(() => {
-				directoryConnector.currentAccount.value = null;
-			});
-			rerender();
-			expect(result.current).toBeNull();
 		});
 	});
 
 	describe('useAuthToken', () => {
 		it('should return the directory connector\'s current auth token information', () => {
-			const { result, rerender } = renderHookWithTestProviders(useAuthToken);
+			const { result } = renderHookWithTestProviders(useAuthToken);
 			expect(result.current).toBe(directoryConnector.authToken.value);
-
-			const newAuthToken = MockAuthToken({
-				username: 'useAuthTokenTest',
-				expires: 95617584000,
-				value: 'eVA8tqM41UJMajBVXJAnOmmODXJIssEN',
-			});
-			act(() => {
-				directoryConnector.authToken.value = newAuthToken;
-			});
-			rerender();
-			expect(result.current).toBe(newAuthToken);
-
-			act(() => {
-				directoryConnector.authToken.value = undefined;
-			});
-			rerender();
-			expect(result.current).toBeUndefined();
 		});
 	});
